+++ examples/simple/settings_bootstrap3_theme_captcha.py
@@ -4,7 +4,7 @@ INSTALLED_APPS = list(INSTALLED_APPS)
 
 try:
     INSTALLED_APPS.append('captcha')
-    INSTALLED_APPS.append('fobi.contrib.plugins.form_elements.fields.captcha')
+    INSTALLED_APPS.append('fobi.contrib.plugins.form_elements.security.captcha')
 except Exception as e:
     pass