+++ examples/simple/settings_bootstrap3_theme_django17_captcha.py
@@ -6,6 +6,6 @@ try:
     INSTALLED_APPS.remove('south') if 'south' in INSTALLED_APPS else None
     INSTALLED_APPS.remove('tinymce') if 'tinymce' in INSTALLED_APPS else None
     INSTALLED_APPS.append('captcha')
-    INSTALLED_APPS.append('fobi.contrib.plugins.form_elements.fields.captcha')
+    INSTALLED_APPS.append('fobi.contrib.plugins.form_elements.security.captcha')
 except Exception as e:
     pass