+++ src/fobi/contrib/plugins/form_elements/fields/date_drop_down/__init__.py
@@ -1,9 +1,9 @@
-__title__ = 'fobi.contrib.plugins.form_elements.fields.birthday'
+__title__ = 'fobi.contrib.plugins.form_elements.fields.date_drop_down'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('default_app_config', 'UID',)
 
-default_app_config = 'fobi.contrib.plugins.form_elements.fields.birthday.apps.Config'
+default_app_config = 'fobi.contrib.plugins.form_elements.fields.date_drop_down.apps.Config'
 
-UID = 'birthday'
+UID = 'date_drop_down'