+++ src/fobi/contrib/plugins/form_elements/security/captcha/__init__.py
@@ -1,9 +1,9 @@
-__title__ = 'fobi.contrib.plugins.form_elements.fields.captcha'
+__title__ = 'fobi.contrib.plugins.form_elements.security.captcha'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
 __all__ = ('default_app_config', 'UID',)
 
-default_app_config = 'fobi.contrib.plugins.form_elements.fields.captcha.apps.Config'
+default_app_config = 'fobi.contrib.plugins.form_elements.security.captcha.apps.Config'
 
 UID = 'captcha'