+++ src/fobi/contrib/plugins/form_elements/security/captcha/forms.py
@@ -1,4 +1,4 @@
-__title__ = 'fobi.contrib.plugins.form_elements.fields.captcha.forms'
+__title__ = 'fobi.contrib.plugins.form_elements.security.captcha.forms'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'