+++ src/fobi/contrib/plugins/form_elements/security/honeypot/apps.py
@@ -1,4 +1,4 @@
-__title__ = 'fobi.contrib.plugins.form_elements.fields.captcha.apps'
+__title__ = 'fobi.contrib.plugins.form_elements.security.honeypot.apps'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
@@ -8,7 +8,7 @@ try:
     from django.apps import AppConfig
 
     class Config(AppConfig):
-        name = label = 'fobi.contrib.plugins.form_elements.fields.captcha'
+        name = label = 'fobi.contrib.plugins.form_elements.security.honeypot'
 
 except ImportError:
     pass