+++ src/fobi/contrib/plugins/form_elements/security/honeypot/defaults.py
@@ -1,6 +1,10 @@
-__title__ = 'django-fobi'
-__version__ = '0.4'
-__build__ = 0x00000f
+__title__ = 'fobi.contrib.plugins.form_elements.security.honeypot.defaults'
 __author__ = 'Artur Barseghyan <artur.barseghyan@gmail.com>'
 __copyright__ = 'Copyright (c) 2014 Artur Barseghyan'
 __license__ = 'GPL 2.0/LGPL 2.1'
+__all__ = (
+    'HONEYPOT_VALUE',
+)
+
+HONEYPOT_VALUE = ''
+