+++ test3.sh
@@ -1,4 +1,4 @@
 reset
 ./uninstall.sh
-./install.sh
+./install3.sh
 python examples/simple/manage.py test fobi --traceback